feat(landingpage): add landingpage desktop
